% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cdf.R
\name{read_cdf}
\alias{read_cdf}
\title{Reads netcdf and returns it as array}
\usage{
read_cdf(filename, nc_header, subset = list())
}
\arguments{
\item{filename}{netcdf file name}

\item{nc_header}{header data, read in from either meta file or data in
netcdf file}

\item{subset}{list object defining which subset of the data to be read}
}
\value{
array with netcdf's data, dim=c(nlon,nlat,bands,steps (months/days),years)
}
\description{
Reads an arbitrary netcdf and returns the values at the lon/lat locations..
Todos:
\itemize{
\item subsetting cells/days/months does not work yet
}
}
